changed backernd structure and added JSON return format
